--- a/bench/brainsIntel_noIPP
+++ b/bench/brainsIntel_noIPP
@@ -1,0 +1,178 @@
+=== cpuinfo ===
+{'arch': 'X86_64',
+ 'arch_string_raw': 'x86_64',
+ 'bits': 64,
+ 'brand_raw': 'Intel(R) Core(TM) i3-7100 CPU @ 3.90GHz',
+ 'count': 4,
+ 'cpuinfo_version': [8, 0, 0],
+ 'cpuinfo_version_string': '8.0.0',
+ 'family': 6,
+ 'flags': ['3dnowprefetch',
+           'abm',
+           'acpi',
+           'adx',
+           'aes',
+           'aperfmperf',
+           'apic',
+           'arat',
+           'arch_perfmon',
+           'art',
+           'avx',
+           'avx2',
+           'bmi1',
+           'bmi2',
+           'bts',
+           'clflush',
+           'clflushopt',
+           'cmov',
+           'constant_tsc',
+           'cpuid',
+           'cpuid_fault',
+           'cx16',
+           'cx8',
+           'de',
+           'ds_cpl',
+           'dtes64',
+           'dtherm',
+           'dts',
+           'ept',
+           'erms',
+           'est',
+           'f16c',
+           'flexpriority',
+           'flush_l1d',
+           'fma',
+           'fpu',
+           'fsgsbase',
+           'fxsr',
+           'ht',
+           'hwp',
+           'hwp_act_window',
+           'hwp_epp',
+           'hwp_notify',
+           'ibpb',
+           'ibrs',
+           'intel_pt',
+           'invpcid',
+           'invpcid_single',
+           'lahf_lm',
+           'lm',
+           'mca',
+           'mce',
+           'md_clear',
+           'mmx',
+           'monitor',
+           'movbe',
+           'mpx',
+           'msr',
+           'mtrr',
+           'nonstop_tsc',
+           'nopl',
+           'nx',
+           'osxsave',
+           'pae',
+           'pat',
+           'pbe',
+           'pcid',
+           'pclmulqdq',
+           'pdcm',
+           'pdpe1gb',
+           'pebs',
+           'pge',
+           'pln',
+           'pni',
+           'popcnt',
+           'pse',
+           'pse36',
+           'pti',
+           'pts',
+           'rdrand',
+           'rdrnd',
+           'rdseed',
+           'rdtscp',
+           'rep_good',
+           'sdbg',
+           'sep',
+           'sgx',
+           'smap',
+           'smep',
+           'ss',
+           'ssbd',
+           'sse',
+           'sse2',
+           'sse4_1',
+           'sse4_2',
+           'ssse3',
+           'stibp',
+           'syscall',
+           'tm',
+           'tm2',
+           'tpr_shadow',
+           'tsc',
+           'tsc_adjust',
+           'tsc_deadline_timer',
+           'tscdeadline',
+           'vme',
+           'vmx',
+           'vnmi',
+           'vpid',
+           'x2apic',
+           'xgetbv1',
+           'xsave',
+           'xsavec',
+           'xsaveopt',
+           'xsaves',
+           'xtopology',
+           'xtpr'],
+ 'hz_actual': [3900041000, 0],
+ 'hz_actual_friendly': '3.9000 GHz',
+ 'hz_advertised': [3900000000, 0],
+ 'hz_advertised_friendly': '3.9000 GHz',
+ 'l1_data_cache_size': 32768,
+ 'l1_instruction_cache_size': 32768,
+ 'l2_cache_associativity': 6,
+ 'l2_cache_line_size': 256,
+ 'l2_cache_size': 262144,
+ 'l3_cache_size': 3145728,
+ 'model': 158,
+ 'python_version': '3.7.7.final.0 (64 bit)',
+ 'stepping': 9,
+ 'vendor_id_raw': 'GenuineIntel'}
+=== plugin config ===
+HDF5PluginBuildOptions(openmp=True, native=True, sse2=True, avx2=True, cpp11=True, filter_file_extension='.so')
+=== Generating test data 0.001 uint16 (2162, 2068) ===
+wrote /dev/shm/bench.h5 testdata 0.894 GB (100, 2162, 2068) 0.391 s 0.008 GB 115.0:1
+OMP_NUM_THREADS 1 read /dev/shm/bench.h5 testdata uint16 0.894 GB 0.551 s 0.572 s 1.623 GB/s
+OMP_NUM_THREADS 2 read /dev/shm/bench.h5 testdata uint16 0.894 GB 0.354 s 0.380 s 2.525 GB/s
+OMP_NUM_THREADS 4 read /dev/shm/bench.h5 testdata uint16 0.894 GB 0.319 s 0.323 s 2.806 GB/s
+TESTREAD read /dev/shm/bench.h5 testdata uint16 0.894 GB 0.320 s 0.423 s 2.799 GB/s
+=== Generating test data 0.001 uint32 (2162, 2068) ===
+wrote /dev/shm/bench.h5 testdata 1.788 GB (100, 2162, 2068) 1.289 s 0.013 GB 137.7:1
+OMP_NUM_THREADS 1 read /dev/shm/bench.h5 testdata uint32 1.788 GB 0.886 s 0.908 s 2.018 GB/s
+OMP_NUM_THREADS 2 read /dev/shm/bench.h5 testdata uint32 1.788 GB 0.599 s 0.626 s 2.987 GB/s
+OMP_NUM_THREADS 4 read /dev/shm/bench.h5 testdata uint32 1.788 GB 0.518 s 0.620 s 3.449 GB/s
+TESTREAD read /dev/shm/bench.h5 testdata uint32 1.788 GB 0.513 s 0.608 s 3.489 GB/s
+=== Generating test data 1 uint16 (2162, 2068) ===
+wrote /dev/shm/bench.h5 testdata 0.894 GB (100, 2162, 2068) 0.365 s 0.135 GB 6.6:1
+OMP_NUM_THREADS 1 read /dev/shm/bench.h5 testdata uint16 0.894 GB 0.604 s 0.615 s 1.481 GB/s
+OMP_NUM_THREADS 2 read /dev/shm/bench.h5 testdata uint16 0.894 GB 0.390 s 0.409 s 2.291 GB/s
+OMP_NUM_THREADS 4 read /dev/shm/bench.h5 testdata uint16 0.894 GB 0.358 s 0.420 s 2.499 GB/s
+TESTREAD read /dev/shm/bench.h5 testdata uint16 0.894 GB 0.357 s 0.449 s 2.503 GB/s
+=== Generating test data 1 uint32 (2162, 2068) ===
+wrote /dev/shm/bench.h5 testdata 1.788 GB (100, 2162, 2068) 1.317 s 0.142 GB 12.6:1
+OMP_NUM_THREADS 1 read /dev/shm/bench.h5 testdata uint32 1.788 GB 0.938 s 0.958 s 1.906 GB/s
+OMP_NUM_THREADS 2 read /dev/shm/bench.h5 testdata uint32 1.788 GB 0.640 s 0.660 s 2.795 GB/s
+OMP_NUM_THREADS 4 read /dev/shm/bench.h5 testdata uint32 1.788 GB 0.556 s 0.648 s 3.216 GB/s
+TESTREAD read /dev/shm/bench.h5 testdata uint32 1.788 GB 0.565 s 0.670 s 3.166 GB/s
+=== Generating test data 1000 uint16 (2162, 2068) ===
+wrote /dev/shm/bench.h5 testdata 0.894 GB (100, 2162, 2068) 0.481 s 0.488 GB 1.8:1
+OMP_NUM_THREADS 1 read /dev/shm/bench.h5 testdata uint16 0.894 GB 0.572 s 0.589 s 1.562 GB/s
+OMP_NUM_THREADS 2 read /dev/shm/bench.h5 testdata uint16 0.894 GB 0.407 s 0.420 s 2.199 GB/s
+OMP_NUM_THREADS 4 read /dev/shm/bench.h5 testdata uint16 0.894 GB 0.386 s 0.422 s 2.314 GB/s
+TESTREAD read /dev/shm/bench.h5 testdata uint16 0.894 GB 0.387 s 0.480 s 2.312 GB/s
+=== Generating test data 1000 uint32 (2162, 2068) ===
+wrote /dev/shm/bench.h5 testdata 1.788 GB (100, 2162, 2068) 1.522 s 0.488 GB 3.7:1
+OMP_NUM_THREADS 1 read /dev/shm/bench.h5 testdata uint32 1.788 GB 1.274 s 1.300 s 1.404 GB/s
+OMP_NUM_THREADS 2 read /dev/shm/bench.h5 testdata uint32 1.788 GB 0.836 s 0.862 s 2.140 GB/s
+OMP_NUM_THREADS 4 read /dev/shm/bench.h5 testdata uint32 1.788 GB 0.696 s 0.794 s 2.571 GB/s
+TESTREAD read /dev/shm/bench.h5 testdata uint32 1.788 GB 0.709 s 0.769 s 2.523 GB/s