Here is a benchmark for issue #144.

You need  Intels IPP to use it (setup.py --ipp=...). At ESRF that can be `module load Intel/2020 ; echo $IPPROOT`, and then unload or get a fresh shell. Or download oneapi from Intel and install it somewhere. You only need IPP, but their compilers and openmp can also be useful. I think IPP also has patches for zfp, zlib, bzip2 as well ?

Gains vary, but look vaguely promising so far (e.g. two computers, seems to reproduce). I have the feeling it is not so good on real data but didn't track down why for the testcase yet. Could be diffraction peaks or something about the mask etc. It might also depend on exact versions of compressors.

Setting up a profile guided build using a benchmark might be interesting, but I would not want to try it via distutils.