@t20100 2021-07-21
Thanks for the investigation!

We've so far avoided to patch the source code of the HDF5 filters, and I would prefer to keep it this way to avoid incompatibilities since there is a few copy of the code here and there (`hdf5plugin` being one more).

Yet, it is still possible to use IPP AND avoid modifying the filter code by compiling it with a `lz4.c` wrapper over IPP and doing the choice of the source file used at compile time.
As long as it is an opt-in (i.e., the user knows what he is doing), it would be fine this way.

Regarding performance gain, I just gave it a try on the `kevlar.h5` dataset available here: http://www.silx.org/pub/pyFAI/pyFAI_UM_2020/data_ID13 (1000 frames: 2167x2070 uint16, 1 chunk per frame) and couldn't see much performance gain, but that's just one benchmark.

Reading the whole `/entry/data/data` dataset at once:
* Without IPP:
  - OMP_NUM_THREADS=1: 7.74 s ± 59.2 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
  - OMP_NUM_THREADS=8: 4.55 s ± 81.1 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
* With IPP:
  - OMP_NUM_THREADS=1: 7.41 s ± 28 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
  - OMP_NUM_THREADS=8: 4.84 s ± 475 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

As for a profiling driven build process, it doesn't seems to worth the effort and it won't cover all tuning anyway: you'll still have to set `OMP_NUM_THREADS` to an appropriate value to get the most out of it.