@jonwright 2021-07-22
OK - I will close this as it does not make sense to merge in the current state. Maybe the upstream code in setup.py can be changed to allow linking against a given lz4 library rather than the one that was vendored, probably better to see with them.